Refactor tests, to only have test that create cluster happen synchronously
