+++ tests/KSail.Tests/Commands/Stop/KSailStopCommandTests.KSailStopHelp_SucceedsAndPrintsIntroductionAndHelp.verified.txt
