+++ tests/KSail.Tests/Commands/Up/KSailUpCommandTests.cs
@@ -2,7 +2,10 @@ using System.CommandLine;
 using System.CommandLine.IO;
 using KSail.Commands.Down;
 using KSail.Commands.Init;
+using KSail.Commands.Start;
+using KSail.Commands.Stop;
 using KSail.Commands.Up;
+using KSail.Commands.Update;
 
 namespace KSail.Tests.Commands.Up;
 
@@ -13,9 +16,28 @@ namespace KSail.Tests.Commands.Up;
 public class KSailUpCommandTests : IAsyncLifetime
 {
   /// <inheritdoc/>
-  public Task DisposeAsync() => Task.CompletedTask;
+  public async Task DisposeAsync() => await CleanupAsync();
   /// <inheritdoc/>
-  public Task InitializeAsync() => Task.CompletedTask;
+  public async Task InitializeAsync() => await CleanupAsync();
+
+  /// <summary>
+  /// Cleanup the test environment.
+  /// </summary>
+  /// <returns></returns>
+  static Task CleanupAsync()
+  {
+    if (Directory.Exists("k8s"))
+      Directory.Delete("k8s", true);
+    if (File.Exists("kind-config.yaml"))
+      File.Delete("kind-config.yaml");
+    if (File.Exists("k3d-config.yaml"))
+      File.Delete("k3d-config.yaml");
+    if (File.Exists("ksail-config.yaml"))
+      File.Delete("ksail-config.yaml");
+    if (!string.IsNullOrEmpty(".sops.yaml"))
+      File.Delete(".sops.yaml");
+    return Task.CompletedTask;
+  }
 
   /// <summary>
   /// Tests that the 'ksail up --help'
@@ -36,171 +58,38 @@ public class KSailUpCommandTests : IAsyncLifetime
   }
 
   /// <summary>
-  /// Tests that a default cluster is created, and the 'ksail up' command is executed successfully.
+  /// Tests that the 'ksail up' command is executed successfully with various configurations.
   /// </summary>
-  [Fact]
-  public async Task KSailUp_WithDefaultOptions_SucceedsAndCreatesDefaultCluster()
+  [Theory]
+  [InlineData("")]
+  [InlineData("-d kind")]
+  [InlineData("--name ksail-advanced-kind --distribution kind --components --sops --post-build-variables --helm-releases")]
+  [InlineData("-d k3d")]
+  [InlineData("--name ksail-advanced-k3d --distribution k3d --components --sops --post-build-variables --helm-releases")]
+  public async Task KSailUp_WithVariousConfigurations_Succeeds(string initArgs)
   {
-    //Cleanup
-    if (Directory.Exists("k8s"))
-      Directory.Delete("k8s", true);
-    File.Delete("kind-config.yaml");
-    File.Delete("ksail-config.yaml");
-
     //Arrange
     var ksailInitCommand = new KSailInitCommand();
     var ksailUpCommand = new KSailUpCommand();
+    var ksailStopCommand = new KSailStopCommand();
+    var ksailStartCommand = new KSailStartCommand();
+    var ksailUpdateCommand = new KSailUpdateCommand();
     var ksailDownCommand = new KSailDownCommand();
 
     //Act
-    int initExitCode = await ksailInitCommand.InvokeAsync("");
+    int initExitCode = await ksailInitCommand.InvokeAsync(initArgs);
     int upExitCode = await ksailUpCommand.InvokeAsync("--destroy");
+    int stopExitCode = await ksailStopCommand.InvokeAsync("");
+    int startExitCode = await ksailStartCommand.InvokeAsync("");
+    int updateExitCode = await ksailUpdateCommand.InvokeAsync("");
     int downExitCode = await ksailDownCommand.InvokeAsync("--registries");
 
     //Assert
     Assert.Equal(0, initExitCode);
     Assert.Equal(0, upExitCode);
+    Assert.Equal(0, stopExitCode);
+    Assert.Equal(0, startExitCode);
+    Assert.Equal(0, updateExitCode);
     Assert.Equal(0, downExitCode);
-
-    //Cleanup
-    Directory.Delete("k8s", true);
-    File.Delete("kind-config.yaml");
-    File.Delete("ksail-config.yaml");
-  }
-
-  /// <summary>
-  /// Tests that a default kind cluster is created, and the 'ksail up' command is executed successfully.
-  /// </summary>
-  [Fact]
-  public async Task KSailUp_WithSimpleKindCluster_SucceedsAndCreatesKindCluster()
-  {
-    //Cleanup
-    if (Directory.Exists("k8s"))
-      Directory.Delete("k8s", true);
-    File.Delete("kind-config.yaml");
-    File.Delete("ksail-config.yaml");
-
-    //Arrange
-    var ksailInitCommand = new KSailInitCommand();
-    var ksailUpCommand = new KSailUpCommand();
-    var ksailDownCommand = new KSailDownCommand();
-
-    //Act
-    int initExitCode = await ksailInitCommand.InvokeAsync("-d kind");
-    int upExitCode = await ksailUpCommand.InvokeAsync("--destroy");
-    int downExitCode = await ksailDownCommand.InvokeAsync("--registries");
-
-    //Assert
-    Assert.Equal(0, initExitCode);
-    Assert.Equal(0, upExitCode);
-    Assert.Equal(0, downExitCode);
-
-    //Cleanup
-    Directory.Delete("k8s", true);
-    File.Delete("kind-config.yaml");
-    File.Delete("ksail-config.yaml");
-  }
-
-  /// <summary>
-  /// Tests that an advanced kind cluster is created, and the 'ksail up' command is executed successfully.
-  /// </summary>
-  [Fact]
-  public async Task KSailUp_WithAdvancedKindCluster_SucceedsAndCreatesAdvancedKindCluster()
-  {
-    //Cleanup
-    if (Directory.Exists("k8s"))
-      Directory.Delete("k8s", true);
-    File.Delete("kind-config.yaml");
-    File.Delete("ksail-config.yaml");
-    File.Delete(".sops.yaml");
-
-    //Arrange
-    var ksailInitCommand = new KSailInitCommand();
-    var ksailUpCommand = new KSailUpCommand();
-    var ksailDownCommand = new KSailDownCommand();
-
-    //Act
-    int initExitCode = await ksailInitCommand.InvokeAsync("--name ksail-advanced-kind --distribution kind --components --sops --post-build-variables --helm-releases");
-    int upExitCode = await ksailUpCommand.InvokeAsync("--destroy");
-    int downExitCode = await ksailDownCommand.InvokeAsync("--registries");
-
-    //Assert
-    Assert.Equal(0, initExitCode);
-    Assert.Equal(0, upExitCode);
-    Assert.Equal(0, downExitCode);
-
-    //Cleanup
-    Directory.Delete("k8s", true);
-    File.Delete("kind-config.yaml");
-    File.Delete("ksail-config.yaml");
-    File.Delete(".sops.yaml");
-  }
-
-  /// <summary>
-  /// Tests that a default k3d cluster is created, and the 'ksail up' command is executed successfully.
-  /// </summary>
-  [Fact]
-  public async Task KSailUp_WithSimpleK3dCluster_SucceedsAndCreatesK3dCluster()
-  {
-    //Cleanup
-    if (Directory.Exists("k8s"))
-      Directory.Delete("k8s", true);
-    File.Delete("k3d-config.yaml");
-    File.Delete("ksail-config.yaml");
-
-    //Arrange
-    var ksailInitCommand = new KSailInitCommand();
-    var ksailUpCommand = new KSailUpCommand();
-    var ksailDownCommand = new KSailDownCommand();
-
-    //Act
-    int initExitCode = await ksailInitCommand.InvokeAsync("-d k3d");
-    int upExitCode = await ksailUpCommand.InvokeAsync("--destroy");
-    int downExitCode = await ksailDownCommand.InvokeAsync("--registries");
-
-    //Assert
-    Assert.Equal(0, initExitCode);
-    Assert.Equal(0, upExitCode);
-    Assert.Equal(0, downExitCode);
-
-    //Cleanup
-    Directory.Delete("k8s", true);
-    File.Delete("k3d-config.yaml");
-    File.Delete("ksail-config.yaml");
-  }
-
-  /// <summary>
-  /// Tests that an advanced k3d cluster is created, and the 'ksail up' command is executed successfully.
-  /// </summary>
-  [Fact]
-  public async Task KSailUp_WithAdvancedK3dCluster_SucceedsAndCreatesAdvancedK3dCluster()
-  {
-    //Cleanup
-    if (Directory.Exists("k8s"))
-      Directory.Delete("k8s", true);
-    File.Delete("k3d-config.yaml");
-    File.Delete("ksail-config.yaml");
-    File.Delete(".sops.yaml");
-
-    //Arrange
-    var ksailInitCommand = new KSailInitCommand();
-    var ksailUpCommand = new KSailUpCommand();
-    var ksailDownCommand = new KSailDownCommand();
-
-    //Act
-    int initExitCode = await ksailInitCommand.InvokeAsync("--name ksail-advanced-k3d --distribution k3d --components --sops --post-build-variables --helm-releases");
-    int upExitCode = await ksailUpCommand.InvokeAsync("--destroy");
-    int downExitCode = await ksailDownCommand.InvokeAsync("--registries");
-
-    //Assert
-    Assert.Equal(0, initExitCode);
-    Assert.Equal(0, upExitCode);
-    Assert.Equal(0, downExitCode);
-
-    //Cleanup
-    Directory.Delete("k8s", true);
-    File.Delete("k3d-config.yaml");
-    File.Delete("ksail-config.yaml");
-    File.Delete(".sops.yaml");
   }
 }
